Reason for this change:
- For now we have added the (childView.getTimeOffset()+animatorChild.getDuration()) as the delay.. This is not correct.
reason:
suppose we try to give an interpolator for each view... The there will be a big imbalance in the animation played from top - bottom.. because of different durations(we expect a same animations for all views).
+
the sort functions doesn't perform as it is expected to, if we are adding up the duration in a cumulative fashion.